Consolidate files into service class
