+++ src/big-query-functions/big-query-service-class.ts
@@ -1,11 +1,58 @@
-import createTable from "./create-table";
-import deleteTable from "./delete-table";
-import insertRowsAsStream from "./insert-rows";
-import tableExists from "./table-exists";
+import { Issue } from "../api-functions/get-issues-by-project";
+import { BigQuery } from "@google-cloud/bigquery";
 
 export default class BQ {
-    static createTable = createTable;
-    static deleteTable = deleteTable;
-    static insertRows = insertRowsAsStream;
-    static tableExists = tableExists;
+    //Creates a new table in the dataset
+    static async createTable(  
+        datasetId: string, // Existing dataset
+        tableId: string, // Table to be created
+        schema: any
+    ) : Promise<void> {
+        const bigquery = new BigQuery();
+    
+        const options = {
+            schema: schema,
+            location: 'US',
+        };
+    
+        const [table] = await bigquery
+            .dataset(datasetId)
+            .createTable(tableId, options);
+        console.log(`Table ${table.id} created`);
+    }
+
+    //Deletes table in the dataset
+    static async deleteTable(datasetId: string, tableId: string): Promise<void> {
+        const bigquery = new BigQuery();
+    
+        await bigquery
+            .dataset(datasetId)
+            .table(tableId)
+            .delete();
+        console.log(`Table ${tableId} deleted`);
+    }
+
+    //Inserts rows into a table
+    static async insertRows(datasetId: string, tableId: string,  rows: Issue[] ): Promise<void> {
+        const bigquery = new BigQuery();
+    
+        await bigquery
+            .dataset(datasetId)
+            .table(tableId)
+            .insert(rows);
+        console.log(`Inserted ${rows.length} rows`);
+    }
+
+    //Returns a promise containing true if the table exists and false if not
+    static async tableExists(datasetId: string, tableId: string) : Promise<boolean> {
+        const bigquery = new BigQuery();
+        const dataset = bigquery.dataset(datasetId);
+    
+        try {
+            await dataset.table(tableId).get();
+            return true;
+        } catch (error) {
+            return false;
+        }
+    }
 }
